MAINT: Move histogram and histogramdd into their own module
800 self-contained lines are easily enough to go in their own file, as are the 500 lines of tests.
For compatibility, the names are still available through `np.lib.function_base.histogram` and `from np.lib.function_base import *`

For simplicity of imports, all of the unqualified `np.` names are now qualified
